# Conflicts:
#	src/main/java/com/depromeet/omobackend/controller/UserController.java
#	src/main/java/com/depromeet/omobackend/service/user/UserService.java
#	src/main/java/com/depromeet/omobackend/service/user/UserServiceImpl.java
#	src/test/java/com/depromeet/omobackend/controller/UserControllerTest.java

+++ src/main/java/com/depromeet/omobackend/dto/response/OmakaseDetailsResponse.java
@@ -29,6 +29,6 @@ public class OmakaseDetailsResponse {
 
     private String businessHours;
 
-    private Integer recommendationCount;
+    private Long recommendationCount;
 
 }
